requirements.txt
# Core dependencies
gradio>=4.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# PDF processing
pypdf>=3.0.0

# Additional utilities
requests>=2.28.0
typing-extensions>=4.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Logging and monitoring
colorlog>=6.0.0